# eb-backend/ip-manager-backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ISP IP Management & Infrastructure API",
    "version": "1.0.0",
    "description": "Production-grade enterprise core IPAM and hardware inventory engine. Fully relational, audited, and protected via authorization matrices."
  },
  "servers": [
    {
      "url": "http://10.249.2.9"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Provide a valid token to access infrastructure components. Obtain this via /api/ip-manager/auth/login"
      }
    },
...
  },
  "paths": {
...
    "/api/ip-manager/clients": {
      "get": {
        "tags": [
          "Clients"
        ],
        "summary": "Get all customer client records",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Clients"
        ],
        "summary": "Create new record in Clients",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created successfully"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/ip-manager/real-ip-subnets": {
      "get": {
        "tags": [
          "Network Subnets"
        ],
        "summary": "Get all allocated Real IP subnets",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Network Subnets"
        ],
        "summary": "Create new record in Network Subnets",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created successfully"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/ip-manager/internal-ip-subnets": {
      "get": {
        "tags": [
          "Network Subnets"
        ],
        "summary": "List Internal Local Allocation Blocks",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Network Subnets"
        ],
        "summary": "Create new record in Network Subnets",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created successfully"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/ip-manager/vlans-v2": {
      "get": {
        "tags": [
          "VLAN Layer 2 Tracking"
        ],
        "summary": "Get all structural VLAN tags",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "VLAN Layer 2 Tracking"
        ],
        "summary": "Create new record in VLAN Layer 2 Tracking",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created successfully"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/ip-manager/wan-v2": {
      "get": {
        "tags": [
          "WAN Branch Configurations"
        ],
        "summary": "Get complete Head-Office to Branch maps",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "WAN Branch Configurations"
        ],
        "summary": "Create new record in WAN Branch Configurations",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created successfully"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/ip-manager/tunnels-v2": {
      "get": {
        "tags": [
          "IP Tunnels & VPNs"
        ],
        "summary": "Get interface tunnels data sets",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "IP Tunnels & VPNs"
        ],
        "summary": "Create new record in IP Tunnels & VPNs",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created successfully"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/ip-manager/vpns": {
      "get": {
        "tags": [
          "IP Tunnels & VPNs"
        ],
        "summary": "Get secure network virtual private listings",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "IP Tunnels & VPNs"
        ],
        "summary": "Create new record in IP Tunnels & VPNs",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created successfully"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/ip-manager/dsp-v2": {
      "get": {
        "tags": [
          "Upstream DSP Providers"
        ],
        "summary": "Get list of circuit carriers",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Upstream DSP Providers"
        ],
        "summary": "Create new record in Upstream DSP Providers",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created successfully"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/ip-manager/credentials-v2": {
      "get": {
        "tags": [
          "Node Access Credentials"
        ],
        "summary": "Fetch core node device keys database matrix",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Node Access Credentials"
        ],
        "summary": "Create new record in Node Access Credentials",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created successfully"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/ip-manager/hardware-assets": {
      "get": {
        "tags": [
          "Hardware Assets"
        ],
        "summary": "Get all hardware assets",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Hardware Assets"
        ],
        "summary": "Create new record in Hardware Assets",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created successfully"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/ip-manager/staff": {
      "get": {
        "tags": [
          "ISP Staff Access"
        ],
        "summary": "Get all staff members",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "ISP Staff Access"
        ],
        "summary": "Create new record in ISP Staff Access",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created successfully"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/ip-manager/users": {
      "get": {
        "summary": "List all IP Manager users",
        "responses": {
          "200": {
            "description": "List of users with permissions"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "post": {
        "summary": "Create a real login user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "role": {
                    "type": "string",
                    "enum": [
                      "admin",
                      "editor",
                      "viewer"
                    ]
                  }
                },
                "required": [
                  "username",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/api/platform/permissions": {
      "get": {
        "summary": "List all permissions",
        "responses": {
          "200": {
            "description": "Full permission catalog"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "default"
        ],
        "summary": "Create a permission",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "permission_name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/platform/roles": {
      "get": {
        "summary": "List all roles",
        "responses": {
          "200": {
            "description": "List of roles"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "post": {
        "summary": "Create a role",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  }
                },
                "required": [
                  "name"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Role created"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/api/platform/roles/{role_id}/permissions": {
      "get": {
        "summary": "Get permissions for Role",
        "tags": [
          "Platform Roles"
        ],
        "parameters": [
          {
            "name": "role_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "summary": "Assign permission to Role",
        "tags": [
          "Platform Roles"
        ],
        "parameters": [
          {
            "name": "role_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "permission_id": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "summary": "Sync all permissions for Role",
        "tags": [
          "Platform Roles"
        ],
        "parameters": [
          {
            "name": "role_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "permission_ids": {
                    "type": "array",
                    "items": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/platform/roles/{role_id}/permissions/{permission_id}": {
      "delete": {
        "summary": "Remove permission from Role",
        "tags": [
          "Platform Roles"
        ],
        "parameters": [
          {
            "name": "role_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "permission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/platform/users/{user_id}/permissions": {
      "get": {
        "summary": "Get permissions for User",
        "tags": [
          "Platform Users"
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "summary": "Assign permission to User",
        "tags": [
          "Platform Users"
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "permission_id": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "summary": "Sync all permissions for User",
        "tags": [
          "Platform Users"
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "permission_ids": {
                    "type": "array",
                    "items": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/platform/users/{user_id}/permissions/{permission_id}": {
      "delete": {
        "summary": "Remove permission from User",
        "tags": [
          "Platform Users"
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "permission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
}